third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-callback-explicit-domain-ALLOW-expected.txt

The child frame will get a function from the parent frame, set the document.domain property and then try to use the function. The function should still be able to access parent frame properties. 

PASS: canGet('window.location.href') should be 'true' and is.
PASS: window.thisIsTopFrame should be 'true' and is.
